Add last_segment_ts to index backfill state (#40428)
Adds `last_segment_ts` to search index backfill state. Restarts backfill if `last_segment_ts` is populated, so that rolling back from the algorithm change is safe.

This isn't technically necessary - changing between old and new algorithms _should_ be fine, but this saves us having to test it and makes it obvious what algorithm was used to create indexes which might be useful.

GitOrigin-RevId: 80bed936e7f974f9e7515479d42b1142afafb3d0