Make node-executor tolerate ZIP directory entries (#40542)
When node-executor downloads a source package, it verifies that the `metadata.json` file is correct by reading all entries from the source ZIP file and comparing them to `modulePaths` in `metadata.json`. However, the ZIP archiver used by node-executor integration tests also creates entries for directories, which makes tests fail if there is a module in a subfolder. This PR fixes node-executor to tolerate ZIP entries for directories, and also adds a subfolder module to tests.

GitOrigin-RevId: ac24d1eb27742620b73a0ae376069470926b40a5