system-udfs: make search filter clauses an array (#40632)
I realized when working on the UI that it’s more convenient for SearchIndexFilters.clauses to be an array rather than an object. In particular, it ensures that the order is stable, and allows users to define multiple filters for the same field (but enable only one of them).

There are no worries about backwards compatibility since the feature isn’t live yet.

GitOrigin-RevId: 2b7f029cded4cc89d83c40a8499300e8243f9eb1