Mikec/token limit mcp logs (#40515)
<img width="1017" height="207" alt="image" src="https://github.com/user-attachments/assets/6c0a3757-85ac-4724-a60d-947d03db76ba" />

https://www.youtube.com/watch?v=1hlDpwePKsg&lc=UgwkBqoiqbmQ6Dj6XSV4AaABAg

This user noted that there is a 25k token limit with Claude code and MCP calls and that its quite frustrating to work with the MCP logs tool currently because of this.

This PR seeks to rectify this by allowing the model to optionally specify a token limit which the code then attempts to estimate.

The model can still supply an optional line limit if it wants to, both limits will be applied and sliced to the nearest entry boundry.

Its currently set to 20k limit to allow for inaccuracies in token estimation.

As mentioned in the `tokens.ts` file I borrowed the estimation logic from this repo: https://github.com/johannschopplich/tokenx I decided not to take a dependency on it and instead just copy it in as its simple and self contained. I dont know what the policy is on this brining in of third party code, but its MIT licensed so I assume its okay?

I have tested this all by running it locally and it all works as expected.

GitOrigin-RevId: a54bab51381988127382be5257a068ff78c7dcfc