Retry persistence writes on database timeout errors (#40938)
Retry all persistence writes on database timeout errors. Currently we just do this for bumping max repeatable timestamp, but we don't want to shut down the whole database on timeout because it just increases load. Instead, we will block the committer until we can connect again.

Changes `Persistence::write` to take slices of `documents` and `indexes` to avoid cloning, which is why this diff looks so big.

GitOrigin-RevId: 0eaff99c240978bc1a521ad59e7e300fad7072b0