Change the meaning of persistence_reader_stream_revision_pairs prev_rev.document being None (#41170)
Currently prev_rev.document is None if the previous revision is a tombstone (`deleted=1`),
and the stream returns an error if the revision does not exist. However, a prev_ts should
never be a tombstone, so turn that into an error. On the other hand, it is possible for the previous
revision to not exist at all - if it has been garbage collected. Normally this should be forbidden by
the retention validator, but the document retention worker intentionally looks at rows out of retention,
and it shouldn't encounter an error.

GitOrigin-RevId: f08ea7ca3b9a8ed9c0fd367dc40ff74bcb58ff29