Move React framework docs (Next.js/TanStack Start) to their own category (#41328)
This moves the Next.js and TanStack Start docs out of the React category and creates new categories. There is now a TanStack category that contains the doc for both TanStack Query and TanStack Start. I also added a short section in the Next.js docs that leads to the React library docs and moved the Next.js server rendering docs above.

GitOrigin-RevId: 9fdf6988ce6a22900bc04387531201658d51060f