convex-backend PR 232: correct link to actions docs (#41525)
Docs currently link to nowhere, I'm mostly confident this is intended to redirect to the actions documentation (similar to the mutations documentation referenced in the previous section).

----

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.

Co-authored-by: Graham Abbott <graham@convex.dev>
GitOrigin-RevId: af82f7e7f16879e1a3e7750a100cb7fb2e10d870