dashboard: automatically reduce page size if paginatedTableDocuments query times out (#41664)
Some Convex users have documents that take longer than 1s to query with the current page size of 25. We'll work to improve query times for these types of documents, but in the meantime, let's have the dashboard gracefully.

This works by adding an error boundary around the data view that looks for the timeout errors. If it detects a timeout error, it sets a local storage item scoped to the deployment, component, and table to a lower page size (currently just halving it) until it finds a page size that can load well for the desired table.

GitOrigin-RevId: 14b14f454810896403bb6216b109482ab3468fb6