Move db index backfill progress writes to IndexWorker from IndexWriter (#41849)
This PR moves the logic for writing index backfill progress to the `IndexWorker` instead of the `IndexWriter`, which is shared across the `IndexWorker` and db cluster migration script. This makes the code cleaner (instead of passing an optional database handle for the `IndexWorker` case, we're only passing the writer an optional channel sender to report progress. This simplifies more complex logic around reading and writing index backfill progress for resumable index backfills.

Also adds a test that index backfill progress gets written to the db.

GitOrigin-RevId: 94ddb196325e31dfaec933f36ab68bc153232c6e