Improve error messages when evaluating convex.config (#42013)
When testing something with components, I realized that the error messages that users get when `convex.config.ts` are obscure. They use internal jargon (“definition file”) instead of the user-facing name (`convex.config.ts`), which can make it hard for the user to decipher which file is incorrect. I updated the error messages to be more user-friendly.

GitOrigin-RevId: 7e06c31aeb7a0cd3c085a70a1ec3255fd35b1950