DB index backfill checkpointing (#42012)
Adds checkpointing to database index backfills so they are resumable (restarting the backend won't erase backfill progress).

We store the `BackfillCursor` in the index backfills table to avoid OCCs in the `_index` table, which only allows one write at a time. This is not consistent with the way search index checkpoints are stored, but we should migrate those out of the `_index` table eventually for the same reason.

GitOrigin-RevId: 1a0b21ac46b3edd1dd52da8a37e1ef173538d235