cli: Ask for confirmation before starting a deploy that deletes large indexes (#42545)
This asks the user for confirmation when deleting large indexes when running `npx convex deploy`:

```
⚠️ This code push will delete the following indexes
from your production deployment (http://127.0.0.1:8000):

⛔ messages.by_body   body, _creationTime  ⚠️ 101,723,891 documents
  → replaced by: messages.by_body   body, author, _creationTime
⛔ messages.by_author   author, _creationTime  (staged)  ⚠️ 101,723,891 documents
⛔ aaa:rateLimits.name   name, key, _creationTime  2 documents
⛔ bbb:emails.by_resendId   resendId, _creationTime  1 document
⛔ bbb:emails.by_status_segment   status, segment, _creationTime  1 document
⛔ bbb:emails.by_finalizedAt   finalizedAt, _creationTime  1 document

The documents that are in the index won’t be deleted, but the index will need
to be backfilled again if you want to restore it later.

? Delete these indexes? (y/N)
```

If the user is running the `deploy` command from a CI, here’s what appears instead:

```
⚠️  This code push will delete the following indexes
from your production deployment (http://127.0.0.1:8000):

⛔ aaa:rateLimits.name   name, key, _creationTime  ⚠️  10,000,000 documents
⛔ bbb:emails.by_finalizedAt   finalizedAt, _creationTime  1 document
⛔ bbb:emails.by_resendId   resendId, _creationTime  1 document
⛔ bbb:emails.by_status_segment   status, segment, _creationTime  1 document

The documents that are in the index won’t be deleted, but the index will need
to be backfilled again if you want to restore it later.

✖ To confirm the push:
• run the deploy command in an interactive terminal
• or run the deploy command with the --allow-deleting-indexes flag
```

This doesn't change the behavior when using `npx convex dev` or deploying to non-prod environments

GitOrigin-RevId: abf47fbd04bb7e8ba962eb4b25aa96922df2f3fb