dashboard: make team selection clearer (#43071)
The current UI in the team/project selector is not totally obvious that you can switch teams. Users get confused all the time. I'd like to add a bit more emphasis to the button that lets you switch teams to alleviate confusion.

This does look a bit weird because the main button in the header looks very similar, but maybe that is fine?

### Before:

<img width="498" height="170" alt="Screenshot 2025-11-13 at 4 28 14 PM" src="https://github.com/user-attachments/assets/8caaead4-f194-414a-917f-ea7cbe6b0ae9" />
<img width="442" height="162" alt="Screenshot 2025-11-13 at 4 28 11 PM" src="https://github.com/user-attachments/assets/bbaf00a3-93e3-4fb2-b94c-f34c3a524358" />

### After:

<img width="451" height="311" alt="Screenshot 2025-11-13 at 4 27 42 PM" src="https://github.com/user-attachments/assets/41cc9254-7f94-44f3-93d5-de605d26cb18" />

<img width="508" height="107" alt="Screenshot 2025-11-13 at 4 27 45 PM" src="https://github.com/user-attachments/assets/53f33da2-bebe-417e-af46-13c5807fd075" />

GitOrigin-RevId: cb4753313fb9bee6ba6ea177a08654393471e9af