Start retention deletion workers right away, instead of waiting for min snapshot ts to advance (#43115)
As written, the deletion worker won't wake up until the min-snapshot cursor is advanced the first time, even if there is lag between the confirmed-deleted cursor and the min-snapshot cursor.
And if we increase the retention delay, the min-snapshot ts will already be far ahead of the desired ts, so we won't advance it for a long time. So the deletion worker won't wake up.

Instead we should always wake up the worker on the first iteration so that we can start working on any lag there may be.

GitOrigin-RevId: bd99ae0ead48e6d438120b1d97bc45d4b6b57673