Move pagination logic from React hook to a wrapper client (#41346)
Copy the reactive forever-expanding reactive paginated query logic from the `usePaginatedQuery` hook into a sidecar/wrapper client, which will eventually enable non-React clients and TanStack Query to use this logic.

A new hook `usePaginatedQuery_future` is now available that is implemented with this new pagination and used in the dashboard to try it out. Today it should be identical, and if that's borne out in the dashboard we can migrate, but we'd also like to modify the behavior in the future by adding options or even changing defaults to improve caching behavior.

GitOrigin-RevId: 64c4cdcdf9b1c8ae21f7a28903140c89db0c0fb5