Apply Convex linter fixes in the monorepo (#43164)
This PR is the result of applying `just format-js` with `@convex-dev/eslint-plugin` added to the monorepo ESLint configuration.

For now, I’ve excluded changes to test files that depend on the old format:
- `npm-packages/udf-tests/convex/analyze.ts`
- `npm-packages/udf-tests/convex/basic.ts`
- `npm-packages/udf-tests/convex/directory/udfs.ts`
- `npm-packages/js-integration-tests/convex/customErrors.ts`
- `npm-packages/js-integration-tests/convex/customErrorsNodeActions.ts`
- `npm-packages/component-tests/projects/basic/convex/componentEntry.ts`
- `npm-packages/component-tests/projects/with-schema/convex/componentEntry.ts`
- `npm-packages/udf-tests/convex/sourceMaps.ts`
- `npm-packages/udf-tests/convex/internal.ts`
- `npm-packages/udf-tests/convex/logging.ts`

I’ll make a subsequent PR that adds specific tests for the old syntax.

This PR also doesn’t add argument validators to functions that need one. I’ll add them manually in a subsequent PR.

GitOrigin-RevId: 8774137e3936a19e27cf28cf0b87dcf0602aea90