node-executor: Include logs in `analyze` error messages (#43582)
When `analyze` fails in isolate, we include the log messages:

```
✖ Error fetching POST  http://127.0.0.1:8000/api/deploy2/start_push 400 Bad Request: InvalidModules: Hit an error while pushing:
Loading the pushed modules encountered the following
    error:
Failed to analyze test.js: Uncaught Error: Uninformative error

'aaa'
'bbb'
    at <anonymous> (../convex/test.ts:4:37)
```

However, we’re not doing the same thing for Node modules:

```
✖ Error fetching POST  http://127.0.0.1:8000/api/deploy2/start_push 400 Bad Request: InvalidModules: Hit an error while pushing:
Loading the pushed modules encountered the following
    error:
Uncaught Failed to analyze test.js: Uninformative error
    at <anonymous> (../convex/test.ts:6:37)
```

This pull request changes the behavior of node-executor so that log messages are also included in the error message. This will be particularly helpful to debug issues with third-party packages that `console.error` + `process.exit` when an API key is missing. (The full support for `process.exit` is coming in #43563.)

GitOrigin-RevId: c3121a1aba3082e9ce00091c36ff1a6533d1d5a7