Fix handling of `process.exit` in Node actions (#43563)
Currently, calling `process.exit` from a Node module will exit the Node executor prematurely. This causes a 500 error, which is confusing to users because they don’t see a helpful log message. In particular, this can be difficult to debug when a third-party dependency calls `process.exit` when an environment variable is missing.

GitOrigin-RevId: 690ce247940cb5e731b67d2d33e461c1876970a0