Log failed queries in the `UdfExecutionLog` (#43937)
Logs query failures to the `UdfExecutionLog`. We already did this for mutations, but we never did it for queries. This was problematic because a customer wasn't seeing this error in their log streams due to the concurrency errors not being logged.

With errors being written to the log, they should be propagated to both the dashboard and any configured log streams.

[This](https://github.com/get-convex/convex/blob/1d2e7308f163e9f5d93f3d29bfb39542eb1568d1/crates/application/src/application_function_runner/mod.rs#L819-L837) is where we do the same thing for mutations.

GitOrigin-RevId: 02fb4b6155e238b7f7e7d0d82c906a8610cab69d