[Text Index Usage Tracking] Start tracking text index ingress (#44038)
Adds new fields in usage tracking for text index ingress + egress. Ingress tracking is now hooked up in the committer and we include only the indexed field using an `estimated_size()` function.

Text index egress is not tracked yet because it requires some additional changes to pipe through the read size from funrun and we have to decide how we want to bill the size of text index reads.

Follow-up items are to:
- [x] charge for text index size
- [ ] start tracking database egress

GitOrigin-RevId: 5c5df769795e78d79c57f7fedb272bb645bba3e5