Put documents in write log for refreshing some system queries (#44313)
Optionally puts `PackedDocument` in the write log for specific tables (hardcoded in a new list).

Motivation: This is useful for refreshing the `DatabaseSnapshotIndexCache` by walking the write log and updating any documents and index keys in the cache. We can then use this up-to-date cache to do efficient queries on the same range, such as for scheduled jobs. (all in a follow up PR, not this one)
GitOrigin-RevId: 2d385a37ea923d524a260d8d8d74633127c5be68