convex-backend PR 277: Docs: Clarified the behavior of message filtering in the query (#44357)
Hey there,

I noticed that the explanation on filtering indexed queries ("matches") does not correspond to the operator used in the example code (`.neq()`).
This also adds a missing closing parenthesis for `.filter(...)` in the same code block.

----

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.

Co-authored-by: Adrian Jagusch <adrian.jagusch@jumax.net>
GitOrigin-RevId: 2f4a9f736cbc3ee6f6a9ed0783e8fdc4e4dc6207