version: freeze the Cursor rules for pre-1.31.0 CLI (#44454)
I’m going to change the Cursor rules soon to add instructions for the new `ctx.db` APIs with explicit table names. We don’t want Cursor to suggest these new APIs if the users haven’t updated to 1.31.0, so I’m going to change the `version` service to return an older version of the rules for old clients.

This PR also adds some tests for the HTTP endpoints with convex-test.

GitOrigin-RevId: 1b61a73132a51ca16acbf3bd7c6bdaeee09f9f85