eslint: add dev dependency on `convex` (#44432)
When we release the `convex` package, Rush updates all peer depencies on `convex` to match the latest version. Since I added a peer depency on `convex` recently in `@convex-dev/eslint-plugin`, this caused the NPM release process to fail. Adding `convex` as a dev dependency fixes the issue.

Ideally we wouldn’t increment the peer dependency version in the ESLint plugin (to still support older `convex` versions), but it’s not a big deal to ask users to update to the latest `convex` version when they want to use a newer ESLint plugin version (and it also avoids the risk of making backwards-incompatible changes in the ESLint plugin).

GitOrigin-RevId: 0f18dba6ff510ce8b642d03cf1d7dea139658f14