Scheduler tolerates deleting scheduled jobs table (#44545)
This fixes an issue where deleting the scheduled functions table would cause the scheduler to fail with `Missing scheduled jobs metadata document`. Now we check if the table is missing and skip the scheduled job if so. If the scheduled job metadata doc is missing but in a table that exists, we throw an error since that shouldn't happen.

GitOrigin-RevId: 26b987c3e43bfb8123f67737da7d949891a0de27