convex-backend PR 268: Optional internal monaco-editor loading (#44537)
Adding optional configuration for the self-hosted dashboard, controllerd by the `NEXT_PUBLIC_LOAD_MONACO_INTERNALLY` environment variable, that allows to load the monaco-editor as a npm package, instaed of the default CDN loader.

This really helpful in cases of network limted deployments that do not have access to the public cdn.jsdeliver.

----

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.

Co-authored-by: FunkMaster-dev <boazgel12018@gmail.com>
GitOrigin-RevId: 5fb887fbbde6cc38d1b61472e9ba3fbf0627f8b5